+++ src/common/FileUploadDropZone/index.tsx
@@ -1,0 +1,2 @@
+export { default as FileUploadDropZone } from "./FileUploadDropZone"
+export type { FileUploadDropZoneProps } from "./FileUploadDropZone"
