+++ src/common/Tree/types.tsx
@@ -1,13 +1,5 @@
-import { ReactNode } from "react"
+import { TreeItem } from "."
 
-export type TreeItem = {
-  id: number
-  parentId: number
-  index: number
-  label?: ReactNode
-  children?: TreeItem[]
-  disabled?: boolean
-}
 export type TreeItemExtended = TreeItem & {
   element?: any
 }
